chat-service/requirements.txt
aio-pika==9.5.5
aiofiles==24.1.0
aiormq==6.8.1
asgiref==3.8.1
attrs==23.1.0
autobahn==24.4.2
Automat==25.4.16
certifi==2023.11.17
cffi==1.16.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
coverage==7.3.3
cryptography==44.0.2
daphne==4.1.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
drf-spectacular==0.27.0
exceptiongroup==1.2.2
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
msgpack==1.1.0
multidict==6.4.3
outcome==1.3.0.post0
packaging==23.2
pamqp==3.3.0
pika==1.2.0
Pillow==10.1.0
pluggy==1.5.0
propcache==0.3.1
protobuf==4.21.12
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pyodbc==5.0.1
pyOpenSSL==25.0.0
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==4.1.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.2.1
referencing==0.32.0
requests==2.31.0
requests-mock==1.11.0
responses==0.24.1
rpds-py==0.13.2
selenium==4.16.0
service-identity==24.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.2
trio-websocket==0.11.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wsproto==1.2.0
yarl==1.20.0
zope.interface==7.2
